+++ src/uci.c
@@ -29,7 +29,7 @@
 
 static pthread_t Timer = 0;
 
-uint64_t uciHash = 2;
+size_t uciHash = 2;
 int64_t uciTimeBuffer = 60;
 bool uciChess960 = false;
 
@@ -45,7 +45,7 @@ static void intro(void)
 {
     uci_puts("id name Demolito " VERSION "\nid author lucasart");
     uci_printf("option name Contempt type spin default %d min -100 max 100\n", Contempt);
-    uci_printf("option name Hash type spin default %" PRIu64 " min 1 max 1048576\n", uciHash);
+    uci_printf("option name Hash type spin default %zu min 1 max 1048576\n", uciHash);
     uci_puts("option name Ponder type check default false");
     uci_printf("option name Threads type spin default %zu min 1 max 256\n", WorkersCount);
     uci_printf("option name Time Buffer type spin default %" PRId64 " min 0 max 1000\n", uciTimeBuffer);
